+++ src/store.js
@@ -61,7 +61,7 @@ normalizeCanvas, contains*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.store = '2021-March-09';
+modules.store = '2021-June-10';
 
 
 // XML_Serializer ///////////////////////////////////////////////////////
@@ -1238,17 +1238,17 @@ SnapSerializer.prototype.loadBlock = function (model, isReporter, object) {
                 model.attributes['var']
             );
         } else {
-        /*
-        // disable JavaScript functions, commented out for now
-        if (model.attributes.s === 'reportJSFunction' &&
-                !Process.prototype.enableJS) {
-            if (window.confirm('enable JavaScript?')) {
-                Process.prototype.enableJS = true;
-            } else {
-                throw new Error('JavaScript is not enabled');
+            // disable loading JavaScript functions
+            /*
+            if (model.attributes.s === 'reportJSFunction' &&
+                    !Process.prototype.enableJS) {
+                if (window.confirm('enable JavaScript?')) {
+                    Process.prototype.enableJS = true;
+                } else {
+                    throw new Error('JavaScript is not enabled');
+                }
             }
-        }
-        */
+            */
             block = SpriteMorph.prototype.blockForSelector(model.attributes.s);
             migration = SpriteMorph.prototype.blockMigrations[
                 model.attributes.s
@@ -1371,7 +1371,6 @@ SnapSerializer.prototype.loadInput = function (model, input, block, object) {
         });
         input.fixLayout();
     } else if (model.tag === 'block' || model.tag === 'custom-block') {
-//        block.silentReplaceInput(input, this.loadBlock(model, true, object));
         block.replaceInput(input, this.loadBlock(model, true, object));
     } else if (model.tag === 'color') {
         input.setColor(this.loadColor(model.contents));